+++ app/Http/Controllers/DishesController.php
@@ -6,23 +6,34 @@ use App\Dishes;
 use App\Repositories\DishRepository\IDishRepository;
 use App\Unit;
 use Illuminate\Http\Request;
-
+use App\Exports\DishExport;
+use Maatwebsite\Excel\Facades\Excel;
+use App\Helper\ICheckAction;
 class DishesController extends Controller
 {
     private $dishesRepository;
+    private $checkAction;
 
-    public function __construct(IDishRepository $dishesRepository)
+    public function __construct(ICheckAction $checkAction, IDishRepository $dishesRepository)
     {
+        $this->checkAction = $checkAction;
         $this->dishesRepository = $dishesRepository;
     }
 
     public function index()
     {
-        $groupmenus = $this->dishesRepository->getGroupMenu();
-        $dishes = Dishes::with('groupMenu.cookArea','unit')->get();
-
-        return view('dishes.index',compact('dishes','groupmenus'));
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleIndex($result);
+        if($check != 0){
+            $groupmenus = $this->dishesRepository->getGroupMenu();
+            $dishes = Dishes::with('material.groupMenu.cookArea','unit')->where('stt','1')->get();
+            $units = $this->dishesRepository->getUnit();
+            return view('dishes.index',compact('dishes','groupmenus','units'));
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
+
     public function viewStore()
     {
         $groupmenus = $this->dishesRepository->getGroupMenu();
@@ -34,35 +45,80 @@ class DishesController extends Controller
 
     public function store(Request $request)
     {
-        $this->dishesRepository->validatorRequestStore($request);
-        return $this->dishesRepository->addDish($request);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleStore($result);
+        if($check != 0){
+            $this->dishesRepository->validatorRequestStore($request);
+            return $this->dishesRepository->addDish($request);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
-    public function viewUpdate($id)
+    public function updateImage(Request $request,$id)
     {
-        $dish = $this->dishesRepository->showUpdateDish($id);
-        $groupmenus = $this->dishesRepository->getGroupMenu();
-        $units = $this->dishesRepository->getUnit();
-        $materials = $this->dishesRepository->getMaterial();
-        return view('dishes.update',compact('dish','groupmenus','units','materials'));
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleUpdate($result);
+        if($check != 0){
+            $this->dishesRepository->validateImage($request);
+            return $this->dishesRepository->updateImageDish($request,$id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
-    public function update(Request $request, $id)
+    public function updateSalePrice(Request $request,$id)
     {
-        // $this->dishesRepository->validatorRequestUpdate($request);
-        return $this->dishesRepository->updateDish($request,$id);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleUpdate($result);
+        if($check != 0){
+            return $this->dishesRepository->updateSalePriceDish($request,$id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
-    public function search(Request $request)
+    public function updateUnit(Request $request,$id)
     {
-        $this->dishesRepository->validatorRequestSearch($request);
-        $dishes = $this->dishesRepository->searchDish($request);
-        $groupmenus = $this->dishesRepository->getGroupMenu();
-        return view('dishes.search',compact('dishes','groupmenus'));
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleUpdate($result);
+        if($check != 0){
+            return $this->dishesRepository->updateUnitDish($request,$id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
+    }
+
+    public function updateStatus(Request $request,$id)
+    {
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleUpdate($result);
+        if($check != 0){
+            return $this->dishesRepository->updateStatusDish($request,$id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
+    }
+
+    public function updateNote(Request $request,$id)
+    {
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleUpdate($result);
+        if($check != 0){
+            return $this->dishesRepository->updateNoteDish($request,$id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 
     public function delete($id)
     {
-        return $this->dishesRepository->deleteDish($id);
+        $result = $this->checkAction->getPermission(auth()->id());
+        $check = $this->dishesRepository->checkRoleDelete($result);
+        if($check != 0){
+            return $this->dishesRepository->deleteDish($id);
+        }else{
+            return view('layouts')->withErrors('Bạn không thuộc quyền truy cập chức năng này');
+        }
     }
 }
